requirements.txt
appdirs==1.4.4
backports.entry-points-selectable==1.1.0
black==21.7b0
certifi==2021.5.30
cfgv==3.3.0
click==8.0.1
distlib==0.3.2
filelock==3.0.12
identify==2.2.13
isort==5.9.3
joblib==1.0.1
mlflow==1.19.0
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.21.1
pandas==1.3.1
pathspec==0.9.0
platformdirs==2.2.0
pre-commit==2.14.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
threadpoolctl==2.2.0
toml==0.10.2
tomli==1.2.1
virtualenv==20.7.2
